utils: make repo safe for "go build ./..."
This directory contains two separate helper programs
that can be run with "go run". It's not possible to cd into
the directory and run "go build", because there are two
different Go source files defining main.main.

The canonical way to indicate this is to put a "// +build ignore"
build tag at the top of such helper programs.
Then they can still be run by

	go run decode_auth.go
	go run test_auth.go

but will not be considered by "cd utils; go build" nor by
commands like "go build ./..." or "go test ./..." in the
root of the repo.